Organize system extensions by system package
